requirements.txt
vnstock == 3.2.5
pandas==2.2.1
requests==2.31.0
bs4==0.0.2
vnai>=2.0
openpyxl==3.1.2
vnstock_ezchart>=0.0.2
fake_useragent==1.5.1
psutil==6.0.0
click==8.1.7
click-repl==0.3.0
seaborn==0.13.2
numpy>=1.23.2,<2.0.0
cryptography==41.0.7  
pyOpenSSL==23.3.0 
python-telegram-bot==20.3
ta==0.11.0
torch==2.9.1